moved nodepropertyspecifier to common
